Created a user controller e2e test file for validating all user end-points
