+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@faker-js/faker": "^9.2.0",
     "@nestjs/common": "^10.0.0",
+    "@nestjs/config": "^3.3.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/mongoose": "^10.1.0",
     "@nestjs/platform-express": "^10.0.0",
